backend/requirements.txt
# Ancient Free Will Database - Backend Requirements

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector database
qdrant-client>=1.15.1

# AI and embeddings
google-generativeai>=0.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.2.0
python-igraph>=0.11
leidenalg>=0.10
python-louvain>=0.16

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9

# Observability & Monitoring
structlog>=23.0.0
prometheus-client>=0.17.0
sentry-sdk[fastapi]>=1.40.0

# Performance & Resilience
pybreaker>=1.0.0
aiohttp>=3.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
